clinker-core: slim PipelineExecutor mod.rs to entry points + re-export root
Move the remaining free helpers out of `executor/mod.rs`, leaving it the
`PipelineExecutor` entry-point impl (run / explain / DAG-walk / route
compile), the trivial `NullStorage` shim, the source-reader registry
aliases, and the submodule re-export surface:

- per-record transform evaluation (`evaluate_single_transform`,
  `evaluate_single_transform_windowed`, `materialize_eval_result`,
  `TransformOutput` / `TransformEvalError`) joins the existing
  `transform.rs`.
- stable-eval-context construction (`build_stable_eval_context`,
  `build_static_vars`, `compute_source_input_arcs`,
  `arcs_reachable_from`) moves to a new `context.rs`.
- the remaining cross-cutting utilities (`scheduled_pass_order`,
  `compute_init_phase_node_set`, `widen_record_to_schema`,
  `recover_engine_stamped_value`, `copy_build_ck_columns`,
  `record_with_emitted_fields`, `single_predecessor`,
  `parse_memory_limit`, `build_arbitrator_from_config`) move to a new
  `util.rs`.

Helpers called only by the executor entry points are widened to
`pub(super)` (`build_stable_eval_context`, `scheduled_pass_order`);
those a dispatch arm consults stay `pub(crate)`; file-local callees
(`materialize_eval_result`, `recover_engine_stamped_value`,
`build_static_vars`, `compute_source_input_arcs`, `arcs_reachable_from`)
become module-private. The `pub(crate)` helpers are re-exported from
`mod.rs`, so every `executor::widen_record_to_schema` /
`executor::single_predecessor` / `executor::compute_init_phase_node_set`
path resolves unchanged. Several doc comments that had drifted away from
their items during the earlier dispatch split are reunited with the
functions they describe.

This drops `executor/mod.rs` from ~4,400 lines to ~1,600 (re-exports
plus the PipelineExecutor impl), with each extracted submodule well
under the 1,200-line target.

Pure relocation: no symbol renames, no public-API change, every existing
test passes unmodified.

Closes #261
