+++ crates/clinker-core/src/executor/mod.rs
@@ -6,6 +6,7 @@ pub mod combine;
 pub(crate) mod combine_dispatch;
 pub(crate) mod commit;
 pub(crate) mod composition_dispatch;
+mod context;
 pub(crate) mod correlation_dispatch;
 pub(crate) mod dispatch;
 mod dlq;
@@ -27,9 +28,11 @@ mod streaming;
 pub(crate) mod time_window;
 pub(crate) mod transform;
 pub(crate) mod transform_dispatch;
+mod util;
 pub(crate) mod watermark;
 pub(crate) mod window_runtime;
 
+use context::build_stable_eval_context;
 pub use dlq::DlqEntry;
 use ingest::{IngestTaskOutcome, ingest_source};
 use params::sum_cpu_io_totals;
@@ -44,27 +47,27 @@ pub(crate) use streaming::{
 use streaming::{compute_streaming_output_specs, streaming_output};
 pub(crate) use transform::CompiledTransform;
 pub use transform::{TransformSpec, build_transform_specs};
+pub(crate) use transform::{evaluate_single_transform, evaluate_single_transform_windowed};
+use util::scheduled_pass_order;
+pub(crate) use util::{
+    build_arbitrator_from_config, compute_init_phase_node_set, copy_build_ck_columns,
+    parse_memory_limit, record_with_emitted_fields, single_predecessor, widen_record_to_schema,
+};
 
 use std::collections::{BTreeMap, HashMap, HashSet};
 use std::io::Read;
 use std::sync::Arc;
 
 use chrono::Utc;
-use clinker_record::{PipelineCounters, Record, RecordStorage, Schema, SchemaBuilder, Value};
+use clinker_record::{PipelineCounters, RecordStorage, Value};
 use indexmap::IndexMap;
 
 use crate::config::PipelineConfig;
 use crate::error::PipelineError;
-use crate::pipeline::arena::Arena;
-use crate::pipeline::index::{GroupByKey, value_to_group_key};
 use crate::pipeline::memory::rss_bytes;
-use crate::pipeline::window_context::PartitionWindowContext;
 use crate::plan::execution::ExecutionPlanDag;
-use cxl::eval::{
-    EvalContext, EvalResult, ProgramEvaluator, SkipReason, StableEvalContext, WallClock,
-};
+use cxl::eval::ProgramEvaluator;
 use cxl::typecheck::Type;
-use petgraph::Direction;
 
 /// Map from source-node name to the input feeding that source.
 ///
@@ -1541,671 +1544,6 @@ impl PipelineExecutor {
             })?;
         Ok((validated_plan.dag().clone(), ()))
     }
-}
-
-/// Resolve the dispatch order for a topological pass via the memory
-/// arbitrator's [`next_runnable`](crate::pipeline::memory::MemoryArbitrator::next_runnable)
-/// instead of walking `topo_order` blindly.
-///
-/// `candidates` is the subset of nodes this pass dispatches — the
-/// init-phase ancestor closure on Pass 1, its complement on Pass 2, or
-/// the whole DAG when there is no init phase. The returned vector lists
-/// those same nodes in the order the executor should dispatch them.
-///
-/// At each step the runnable frontier is every not-yet-emitted candidate
-/// whose graph predecessors *within `candidates`* are all emitted. The
-/// arbitrator picks one frontier node by predicted memory impact (headroom
-/// fit, then largest immediate freed-on-complete, then largest downstream
-/// subtree reclaim, then stable topo index); finishing the chosen node may
-/// open new frontier members for the next step.
-///
-/// Restricting the predecessor check to `candidates` is what lets Pass 1
-/// run before Pass 2: an init-phase node is gated only by its init-phase
-/// ancestors, never by runtime-DAG nodes that have not been dispatched
-/// yet. The init-phase closure is exactly the set of init Transforms plus
-/// their transitive upstream ancestors, so every init node's
-/// non-init predecessors are themselves impossible — the restriction is
-/// total within the closure, not a relaxation of any real dependency.
-///
-/// Determinism: with no volume estimates every candidate's
-/// `predicted_peak_bytes` is `0`, so the arbitrator's headroom, freed-bytes,
-/// and subtree-reclaim tiers are all-equal and selection collapses to the
-/// lowest stable topo index. Greedily taking the lowest-topo-index frontier node
-/// reproduces a plain front-to-back walk of `topo_order` byte-for-byte —
-/// scheduling reorders the dispatch sequence only when real estimates
-/// distinguish the candidates, and never changes record output, which is
-/// independent of dispatch order.
-fn scheduled_pass_order(
-    plan: &crate::plan::execution::ExecutionPlanDag,
-    arbitrator: &crate::pipeline::memory::MemoryArbitrator,
-    candidates: &HashSet<petgraph::graph::NodeIndex>,
-) -> Vec<petgraph::graph::NodeIndex> {
-    use crate::pipeline::memory::SchedulingHint;
-
-    let mut emitted: HashSet<petgraph::graph::NodeIndex> = HashSet::with_capacity(candidates.len());
-    let mut order: Vec<petgraph::graph::NodeIndex> = Vec::with_capacity(candidates.len());
-    let hint: &dyn SchedulingHint = plan;
-
-    // A node is runnable once every in-`candidates` predecessor is
-    // already emitted. Recomputed each step from `emitted` so that
-    // finishing a node exposes its now-unblocked successors.
-    while order.len() < candidates.len() {
-        let runnable: Vec<petgraph::graph::NodeIndex> = candidates
-            .iter()
-            .copied()
-            .filter(|idx| !emitted.contains(idx))
-            .filter(|&idx| {
-                plan.graph
-                    .neighbors_directed(idx, petgraph::Direction::Incoming)
-                    .all(|pred| !candidates.contains(&pred) || emitted.contains(&pred))
-            })
-            .collect();
-
-        // The candidate set is the node set of an acyclic graph, so as
-        // long as work remains at least one node has all in-set
-        // predecessors satisfied. An empty frontier with work left would
-        // mean a cycle, which the planner forbids.
-        debug_assert!(
-            !runnable.is_empty(),
-            "scheduled_pass_order: empty runnable frontier with {} candidates left to emit",
-            candidates.len() - order.len()
-        );
-
-        let chosen = arbitrator.next_runnable(&runnable, hint);
-        emitted.insert(chosen);
-        order.push(chosen);
-    }
-
-    order
-}
-
-/// Build the pipeline-stable evaluation context.
-///
-/// Called ONCE per pipeline run at the top of `execute_dag_branching`. The
-/// returned `StableEvalContext` is reused (via borrow) at every per-record
-/// dispatch site, killing the prior `String::clone` + `IndexMap::clone`
-/// Compute the init-phase ancestor closure for a compiled plan.
-///
-/// Walks every `PlanNode::Transform` whose payload phase is
-/// [`crate::config::Phase::Init`] and unions their transitive
-/// upstream ancestors via reverse-BFS along the graph's incoming
-/// edges. The init-phase transforms themselves are included.
-///
-/// Returns an empty set when no init-phase transforms exist — the
-/// caller then falls through to a single-pass topo walk.
-///
-/// E164 validation guarantees init-phase Transforms are terminal
-/// (no runtime descendants), so the closure forms a well-bounded
-/// init sub-DAG that Pass 1 can run to completion before Pass 2
-/// starts.
-pub(crate) fn compute_init_phase_node_set(
-    plan: &crate::plan::execution::ExecutionPlanDag,
-) -> std::collections::HashSet<petgraph::graph::NodeIndex> {
-    use crate::config::Phase as ConfPhase;
-    use crate::plan::execution::PlanNode;
-    let mut set: std::collections::HashSet<petgraph::graph::NodeIndex> =
-        std::collections::HashSet::new();
-    let mut stack: Vec<petgraph::graph::NodeIndex> = plan
-        .graph
-        .node_indices()
-        .filter(|&idx| match &plan.graph[idx] {
-            PlanNode::Transform {
-                resolved: Some(p), ..
-            } => p.phase == ConfPhase::Init,
-            _ => false,
-        })
-        .collect();
-    while let Some(idx) = stack.pop() {
-        if !set.insert(idx) {
-            continue;
-        }
-        for parent in plan
-            .graph
-            .neighbors_directed(idx, petgraph::Direction::Incoming)
-        {
-            stack.push(parent);
-        }
-    }
-    set
-}
-
-/// per-record allocation profile. `pipeline_start_time` must be frozen at
-/// pipeline start so `$pipeline.start_time` is deterministic within a run.
-/// The `now` keyword uses `ctx.stable.clock.now()` (wall-clock) and is
-/// intentionally non-deterministic.
-fn build_stable_eval_context(
-    config: &PipelineConfig,
-    pipeline_start_time: chrono::NaiveDateTime,
-    execution_id: &str,
-    batch_id: &str,
-    pipeline_vars: &IndexMap<String, Value>,
-    static_vars_overrides: &IndexMap<String, Value>,
-) -> StableEvalContext {
-    // Seed pipeline-scope vars with declared defaults from every
-    // Transform's `declares:` entries; runtime injections (channel
-    // overrides, test seeds) overlay on top.
-    let mut seeded = crate::config::collect_pipeline_var_defaults(&config.nodes);
-    for (k, v) in pipeline_vars {
-        seeded.insert(k.clone(), v.clone());
-    }
-    StableEvalContext {
-        clock: Box::new(WallClock),
-        pipeline_start_time,
-        pipeline_name: Arc::from(config.pipeline.name.as_str()),
-        pipeline_execution_id: Arc::from(execution_id),
-        pipeline_batch_id: Arc::from(batch_id),
-        pipeline_counters: PipelineCounters::default(),
-        pipeline_vars: Arc::new(std::sync::RwLock::new(seeded)),
-        source_vars: Arc::new(std::sync::RwLock::new(std::collections::HashMap::new())),
-        source_input_arcs: Arc::new(compute_source_input_arcs(config)),
-        static_vars: Arc::new(build_static_vars(config, static_vars_overrides)),
-    }
-}
-
-/// Build the `$vars.*` runtime value map from the pipeline's `vars:`
-/// block, with `overrides` (channel-supplied) layered on top. Channel
-/// adds extend the map; channel overrides replace.
-fn build_static_vars(
-    config: &PipelineConfig,
-    overrides: &IndexMap<String, Value>,
-) -> IndexMap<String, Value> {
-    let mut out = config
-        .pipeline
-        .vars
-        .as_ref()
-        .map(crate::config::convert_vars)
-        .unwrap_or_default();
-    for (k, v) in overrides {
-        out.insert(k.clone(), v.clone());
-    }
-    out
-}
-
-/// Walk the YAML config to map every `Merge` / `Combine` named input to
-/// the source-file `Arc<str>`s of the upstream `Source` node(s) it
-/// transitively reads from. Used by `Expr::QualifiedSourceAccess` eval
-/// to look up `source_vars` entries written by upstream source-scope
-/// Transform writers (qualified post-merge `$source.<input>.<key>` reads).
-///
-/// For Merge, each entry in `inputs:` becomes its own input name (the
-/// referenced node name itself, since Merge does not rename). For
-/// Combine, the IndexMap key is the input name. Walk-back follows the
-/// consumer-side `input:` field through Transform/Aggregate/Route nodes
-/// until a Source is reached; nested Merges fan out, collecting every
-/// reachable Source's path Arc.
-fn compute_source_input_arcs(
-    config: &PipelineConfig,
-) -> std::collections::HashMap<String, Vec<Arc<str>>> {
-    use crate::config::pipeline_node::PipelineNode;
-    let mut by_name: HashMap<&str, &PipelineNode> = HashMap::new();
-    for node in &config.nodes {
-        by_name.insert(node.value.name(), &node.value);
-    }
-    let mut out: std::collections::HashMap<String, Vec<Arc<str>>> =
-        std::collections::HashMap::new();
-    for node in &config.nodes {
-        match &node.value {
-            PipelineNode::Combine { header, .. } => {
-                for (input_name, upstream) in header.input.iter() {
-                    let arcs = arcs_reachable_from(&by_name, upstream.value.name());
-                    if !arcs.is_empty() {
-                        out.entry(input_name.clone()).or_default().extend(arcs);
-                    }
-                }
-            }
-            PipelineNode::Merge { header, .. } => {
-                for upstream in &header.inputs {
-                    let upstream_name = upstream.value.name();
-                    let arcs = arcs_reachable_from(&by_name, upstream_name);
-                    if !arcs.is_empty() {
-                        out.entry(upstream_name.to_string())
-                            .or_default()
-                            .extend(arcs);
-                    }
-                }
-            }
-            _ => {}
-        }
-    }
-    for arcs in out.values_mut() {
-        arcs.sort();
-        arcs.dedup();
-    }
-    out
-}
-
-fn arcs_reachable_from(
-    by_name: &HashMap<&str, &crate::config::pipeline_node::PipelineNode>,
-    start: &str,
-) -> Vec<Arc<str>> {
-    use crate::config::pipeline_node::PipelineNode;
-    let mut out = Vec::new();
-    let mut stack = vec![start.to_string()];
-    let mut seen: HashSet<String> = HashSet::new();
-    while let Some(name) = stack.pop() {
-        if !seen.insert(name.clone()) {
-            continue;
-        }
-        let Some(node) = by_name.get(name.as_str()) else {
-            continue;
-        };
-        match node {
-            PipelineNode::Source { config: body, .. } => {
-                let path = body.source.path_str();
-                if !path.is_empty() {
-                    out.push(Arc::from(path));
-                }
-            }
-            PipelineNode::Merge { header, .. } => {
-                for upstream in &header.inputs {
-                    stack.push(upstream.value.name().to_string());
-                }
-            }
-            PipelineNode::Combine { header, .. } => {
-                for upstream in header.input.values() {
-                    stack.push(upstream.value.name().to_string());
-                }
-            }
-            other => {
-                if let Some(input) = other.input_node_name() {
-                    stack.push(input.to_string());
-                }
-            }
-        }
-    }
-    out
-}
-
-/// Evaluate a single transform against a record, returning the
-/// modified record with emitted fields and metadata merged in.
-///
-/// Used by the DAG-walking executor for per-node evaluation. The
-/// record schema is widened in-place (rebuilt onto a schema that
-/// includes every emitted field) so `Record::set` always hits a
-/// known slot. Downstream positional-index consumers (aggregation
-/// `group_by_indices`) rely on the upstream layout, so the widened
-/// schema preserves every upstream column at its original index.
-///
-/// Single emitted (or skipped) output of a per-record transform
-/// evaluation. The `Result<(), SkipReason>` distinguishes a real emit
-/// (`Ok(())`) from a `filter`/`distinct` skip; the leading `Record` is
-/// the output record (or the original record on Skip). One entry per
-/// emitted record under `emit each` fan-out.
-pub(crate) type TransformOutput = (Record, Result<(), SkipReason>);
-
-/// Error shape returned by [`evaluate_single_transform`] and
-/// [`evaluate_single_transform_windowed`] — pairs the transform name
-/// with the underlying [`cxl::eval::EvalError`] for DLQ classification.
-pub(crate) type TransformEvalError = (String, cxl::eval::EvalError);
-
-/// Returns a `Vec` of `(record, Ok | Skip)` entries — typically one
-/// entry per call, but `emit each` fan-out produces N entries from a
-/// single input record. On error, returns `(transform_name, EvalError)`.
-#[allow(clippy::result_large_err)]
-pub(crate) fn evaluate_single_transform(
-    record: &Record,
-    transform_name: &str,
-    evaluator: &mut ProgramEvaluator,
-    ctx: &EvalContext,
-    _output_schema: &Arc<Schema>,
-) -> Result<Vec<TransformOutput>, TransformEvalError> {
-    let input = record;
-    let result = evaluator
-        .eval_record::<NullStorage>(ctx, input, None)
-        .map_err(|e| (transform_name.to_string(), e))?;
-    Ok(materialize_eval_result(input, result))
-}
-
-/// Convert an [`EvalResult`] into the per-record dispatch shape used
-/// by the executor: one `(Record, Ok)` per emitted body record, or a
-/// single `(Record, Skip)` entry on Filtered/Duplicate. The fan-out
-/// case applies emitted fields and record-var writes on a fresh
-/// `record_with_emitted_fields` projection of `input` once per
-/// `EmitOne`; downstream cursor advancement happens once per emitted
-/// record at the call site.
-fn materialize_eval_result(input: &Record, result: EvalResult) -> Vec<TransformOutput> {
-    match result {
-        EvalResult::Emit {
-            fields: emitted,
-            record_vars,
-            ..
-        } => {
-            let mut out = record_with_emitted_fields(input, &emitted);
-            for (name, value) in &emitted {
-                out.set(name, value.clone());
-            }
-            for (key, value) in *record_vars {
-                let _ = out.set_record_var(&key, value);
-            }
-            vec![(out, Ok(()))]
-        }
-        EvalResult::EmitMany { records } => {
-            let mut acc = Vec::with_capacity(records.len());
-            for rec in records {
-                let cxl::eval::EmitOne {
-                    fields: emitted,
-                    record_vars,
-                } = rec;
-                let mut out = record_with_emitted_fields(input, &emitted);
-                for (name, value) in &emitted {
-                    out.set(name, value.clone());
-                }
-                for (key, value) in *record_vars {
-                    let _ = out.set_record_var(&key, value);
-                }
-                acc.push((out, Ok(())));
-            }
-            acc
-        }
-        EvalResult::Skip(reason) => vec![(input.clone(), Err(reason))],
-    }
-}
-
-/// Same as [`evaluate_single_transform`] but threads a
-/// [`PartitionWindowContext`] derived from the resolved
-/// [`crate::executor::window_runtime::WindowRuntime`] so `$window.*`
-/// expressions resolve against the transform's analytic window.
-/// `window_index` indexes `plan.indices_to_build`.
-///
-/// The caller resolves `runtime` via
-/// `ctx.window_runtime.resolve(&spec.root, window_index)` and is
-/// responsible for surfacing `PipelineError::Internal` if `resolve`
-/// returns `None` — at this point in the dispatch arm the upstream
-/// operator has already finalized its emit, so a missing runtime is
-/// always an invariant violation.
-///
-/// `record_pos` is the record's position in the arena, derived by the
-/// caller from the spec's root: `(rn - 1)` for `Source(_)` roots
-/// (Phase-0 materializes source records in row-number order), or the
-/// per-record enumerate index for `Node{..}` and `ParentNode{..}` roots
-/// (the upstream operator's emit order is the arena's order).
-///
-/// If the record's group-by values do not resolve to a partition (null
-/// in any group key, or no matching partition), the evaluator runs
-/// without a window context — matching the legacy inline arena path's
-/// behavior.
-#[allow(clippy::too_many_arguments, clippy::result_large_err)]
-pub(crate) fn evaluate_single_transform_windowed(
-    record: &Record,
-    transform_name: &str,
-    evaluator: &mut ProgramEvaluator,
-    ctx: &EvalContext,
-    plan: &ExecutionPlanDag,
-    window_index: usize,
-    runtime: &crate::executor::window_runtime::WindowRuntime,
-    record_pos: u64,
-) -> Result<Vec<TransformOutput>, TransformEvalError> {
-    let spec = &plan.indices_to_build[window_index];
-    let arena = runtime.arena.as_ref();
-    let index = runtime.index.as_ref();
-
-    let key: Option<Vec<GroupByKey>> = spec
-        .group_by
-        .iter()
-        .map(|field| {
-            let val = record.get(field).cloned().unwrap_or(Value::Null);
-            value_to_group_key(&val, field, None, record_pos)
-                .ok()
-                .flatten()
-        })
-        .collect();
-
-    let sort_fields: Vec<&str> = spec.sort_by.iter().map(|s| s.field.as_str()).collect();
-    let result = if let Some(key) = key {
-        if let Some(partition) = index.get(&key) {
-            let pos_in_partition = partition.iter().position(|&p| p == record_pos).unwrap_or(0);
-            let wctx =
-                PartitionWindowContext::new(arena, partition, pos_in_partition, &sort_fields);
-            evaluator
-                .eval_record(ctx, record, Some(&wctx))
-                .map_err(|e| (transform_name.to_string(), e))?
-        } else {
-            evaluator
-                .eval_record::<Arena>(ctx, record, None)
-                .map_err(|e| (transform_name.to_string(), e))?
-        }
-    } else {
-        evaluator
-            .eval_record::<Arena>(ctx, record, None)
-            .map_err(|e| (transform_name.to_string(), e))?
-    };
-
-    Ok(materialize_eval_result(record, result))
-}
-
-/// Re-project `input` onto `target`: allocate a fresh `Record` whose
-/// schema is `target`, copying over any upstream field that `target`
-/// still declares. Used at operator boundaries to canonicalize the
-/// `Arc<Schema>` on the Record so downstream `Arc::ptr_eq` checks hit
-/// the fast path.
-pub(crate) fn widen_record_to_schema(input: &Record, target: &Arc<Schema>) -> Record {
-    if Arc::ptr_eq(input.schema(), target) {
-        return input.clone();
-    }
-    let mut values: Vec<clinker_record::Value> = Vec::with_capacity(target.column_count());
-    for (i, col) in target.columns().iter().enumerate() {
-        let v = match input.get(col.as_ref()) {
-            Some(v) => v.clone(),
-            None => recover_engine_stamped_value(input, target, i, col.as_ref()),
-        };
-        values.push(v);
-    }
-    let mut out = Record::new(Arc::clone(target), values);
-    // Widening reshapes the schema for the same document's row; carry
-    // the envelope context forward so a downstream node reading
-    // `$doc.<section>.<field>` resolves against the originating
-    // document rather than the empty synthetic context.
-    out.set_doc_ctx(Arc::clone(input.doc_ctx()));
-    out
-}
-
-/// Recover an engine-stamped column's value when the input does not
-/// expose it under the target column name. The N-ary combine
-/// decomposition encodes intermediate-step columns as
-/// `__<qualifier>__<name>`, so a `$ck.<field>` snapshot column
-/// arrives at the final step's body construction under
-/// `__<driver>__$ck.<field>` rather than `$ck.<field>` directly.
-/// Walk the input schema for any column ending in `__<col>` and
-/// recover the value; otherwise default to `Value::Null` (the column
-/// remains an unstamped slot, matching the behavior for build sources
-/// that never declared the correlation-key field).
-fn recover_engine_stamped_value(
-    input: &Record,
-    target: &Arc<Schema>,
-    target_idx: usize,
-    target_col: &str,
-) -> clinker_record::Value {
-    let is_engine_stamped = target
-        .field_metadata(target_idx)
-        .is_some_and(|m| m.is_engine_stamped());
-    if !is_engine_stamped {
-        return clinker_record::Value::Null;
-    }
-    let suffix = format!("__{target_col}");
-    if let Some(j) = input
-        .schema()
-        .columns()
-        .iter()
-        .position(|n| n.as_ref().ends_with(&suffix))
-    {
-        return input.values()[j].clone();
-    }
-    clinker_record::Value::Null
-}
-
-/// Copy build-side `$ck.<field>` columns from `build` into `out` per
-/// the combine's `propagate_ck` spec. Driver wins on collision: if
-/// `out` already carries a non-null value for the column (because
-/// `widen_record_to_schema(driver, …)` filled it), the build value is
-/// not written. This is the single runtime CK-copy path used by every
-/// combine strategy (HashBuildProbe, IEJoin, GraceHash, SortMerge);
-/// keeping the policy here means a strategy never has to encode the
-/// `$ck` semantics itself.
-///
-/// `Driver` copies only the engine-managed synthetic CK
-/// (`$ck.aggregate.<name>`, stamped
-/// [`clinker_record::FieldMetadata::AggregateGroupIndex`]) — the user-
-/// declared source CK is suppressed. `All` copies every `$ck.*` column
-/// the build record carries. `Named(set)` copies only the listed
-/// source-CK fields plus every synthetic-CK column unconditionally.
-///
-/// Synthetic CK is engine-managed lineage from a relaxed aggregate to
-/// its source rows; the user did not declare it, so `propagate_ck`
-/// (a knob over user-declared source CK) has no semantic meaning over
-/// it. Without this exemption, the detect-phase fan-out from a
-/// downstream failure to the aggregator's per-group source-row table
-/// would lose its bridge whenever a Combine sat between the relaxed
-/// aggregate and the failing node.
-pub(crate) fn copy_build_ck_columns(
-    out: &mut Record,
-    build: &Record,
-    spec: &crate::config::pipeline_node::PropagateCkSpec,
-) {
-    use crate::config::pipeline_node::PropagateCkSpec;
-    let build_schema = Arc::clone(build.schema());
-    for (idx, col) in build_schema.columns().iter().enumerate() {
-        let Some(field_name) = col.strip_prefix("$ck.") else {
-            continue;
-        };
-        let is_synthetic = matches!(
-            build_schema.field_metadata(idx),
-            Some(clinker_record::FieldMetadata::AggregateGroupIndex { .. }),
-        );
-        let allowed = is_synthetic
-            || match spec {
-                PropagateCkSpec::Driver => false,
-                PropagateCkSpec::All => true,
-                PropagateCkSpec::Named(names) => names.contains(field_name),
-            };
-        if !allowed {
-            continue;
-        }
-        if out.schema().index(col.as_ref()).is_none() {
-            // Output schema didn't widen for this column — the
-            // plan-time `combine_output_row` filtered it out
-            // already. Skip rather than silently lose data.
-            continue;
-        }
-        // Driver wins on collision: a non-null value at this slot
-        // came from the driver via `widen_record_to_schema`. Only
-        // fill when the slot is still null. Synthetic-CK names are
-        // unique per aggregate (`$ck.aggregate.<name>`), so a driver
-        // and build pairing can collide only when both descend from
-        // the same relaxed aggregate, in which case the slot is
-        // already populated with the correct lineage.
-        match out.get(col.as_ref()) {
-            Some(clinker_record::Value::Null) | None => {
-                let v = &build.values()[idx];
-                out.set(col.as_ref(), v.clone());
-            }
-            Some(_) => {}
-        }
-    }
-}
-
-/// Widen `input`'s schema in place to include every key in `emitted`
-/// that is not already declared. Allocates a fresh `Arc<Schema>` only
-/// when new names appear; otherwise clones `input`. Used by the legacy
-/// linear pipeline path where the emit set is determined at eval time
-/// rather than via a plan-time `output_schema`, and by the
-/// commit-phase window recompute which materializes new output rows
-/// against the same emit shape the dispatcher's window arm produced.
-pub(crate) fn record_with_emitted_fields(
-    input: &Record,
-    emitted: &IndexMap<String, clinker_record::Value>,
-) -> Record {
-    let mut missing: Vec<&str> = Vec::new();
-    for key in emitted.keys() {
-        if input.schema().index(key).is_none() {
-            missing.push(key.as_str());
-        }
-    }
-    if missing.is_empty() {
-        return input.clone();
-    }
-    let n = input.schema().column_count();
-    let mut builder = SchemaBuilder::with_capacity(n + missing.len());
-    // Preserve every existing column AND its `FieldMetadata`. Without
-    // explicit forwarding the engine-stamp annotation on `$ck.<field>`
-    // shadow columns is dropped and downstream readers (the buffer-key
-    // extractor at the Output arm, the projection fast path) treat the
-    // column as a user-declared one.
-    for i in 0..n {
-        let name = input
-            .schema()
-            .column_name(i)
-            .expect("column_name within column_count");
-        match input.schema().field_metadata(i) {
-            Some(meta) => builder = builder.with_field_meta(name, meta.clone()),
-            None => builder = builder.with_field(name),
-        }
-    }
-    for name in missing {
-        builder = builder.with_field(name);
-    }
-    let widened = builder.build();
-    widen_record_to_schema(input, &widened)
-}
-
-/// Parse memory limit from config (default 512MB).
-/// Plan-invariant predecessor lookup for nodes that require exactly one
-/// upstream input (currently `PlanNode::Aggregation`). Returns
-/// `PipelineError::Internal` on misshapen plans — never panics. Mirrors
-/// DataFusion's `internal_err!` macro.
-pub(crate) fn single_predecessor(
-    plan: &ExecutionPlanDag,
-    node_idx: petgraph::graph::NodeIndex,
-    op: &'static str,
-    node_name: &str,
-) -> Result<petgraph::graph::NodeIndex, PipelineError> {
-    let preds: Vec<_> = plan
-        .graph
-        .neighbors_directed(node_idx, Direction::Incoming)
-        .collect();
-    match preds.as_slice() {
-        [p] => Ok(*p),
-        _ => Err(PipelineError::Internal {
-            op,
-            node: node_name.to_string(),
-            detail: format!("expected exactly 1 predecessor, got {}", preds.len()),
-        }),
-    }
-}
-
-pub(crate) fn parse_memory_limit(config: &PipelineConfig) -> usize {
-    config
-        .pipeline
-        .memory
-        .limit
-        .as_ref()
-        .and_then(|s| {
-            let s = s.trim();
-            if let Some(num) = s.strip_suffix('G').or_else(|| s.strip_suffix('g')) {
-                num.parse::<usize>().ok().map(|n| n * 1024 * 1024 * 1024)
-            } else if let Some(num) = s.strip_suffix('M').or_else(|| s.strip_suffix('m')) {
-                num.parse::<usize>().ok().map(|n| n * 1024 * 1024)
-            } else {
-                s.parse::<usize>().ok()
-            }
-        })
-        .unwrap_or(512 * 1024 * 1024) // 512MB default
-}
-
-/// Build a `MemoryArbitrator` from the pipeline-level `memory:` block.
-/// Resolves `memory.limit` through `parse_memory_limit_bytes` (defaults
-/// to 512 MiB when omitted) and chooses the active policy via
-/// `BackpressureKnob::build_policy`. Used by both the pipeline-scoped
-/// arbitrator and every per-arm budget the dispatch path constructs.
-pub(crate) fn build_arbitrator_from_config(
-    config: &PipelineConfig,
-) -> crate::pipeline::memory::MemoryArbitrator {
-    let mem = &config.pipeline.memory;
-    let limit = crate::pipeline::memory::parse_memory_limit_bytes(mem.limit.as_deref());
-    crate::pipeline::memory::MemoryArbitrator::with_policy(
-        limit,
-        0.80,
-        mem.backpressure.build_policy(),
-    )
 }
 
 #[cfg(test)]
